+++ r/R/lcg.R
@@ -4,27 +4,22 @@
 #' An integer in `{0, 1, ..., modulus-1}`, if `standardize = FALSE`,
 #' or a numeric in `[0,1)`, if `standardize = TRUE`,
 #'
+#' @details
+#' The LCG parameters are $modulus = 2^16+1$, $a = 75$, and $c = 74$,
+#' which are small enough to to be handled by Bash.  They also happens to be
+#' the one used by the fabolous Sinclair ZX81.
+#'
 #' @reference
 #' https://en.wikipedia.org/wiki/Linear_congruential_generator
 lcg <- local({
+  ## LCG parameters
+  modulus <- 2^16+1
+  a <- 75
+  c <- 74
+  
   .seed <- NULL
   
-  function(modulus = getOption("lcg.params")[["modulus"]], a = getOption("lcg.params")[["a"]], c = getOption("lcg.params")[["c"]], seed = NULL) {
-    ## Set default LCG parameters, if not already set
-    if (is.null(getOption("lcg.params"))) {
-      lcg_set_params()
-    }
-    
-    if (is.null(modulus)) modulus <- getOption("lcg.params")[["modulus"]]
-    if (is.null(a)) a <- getOption("lcg.params")[["a"]]
-    if (is.null(c)) c <- getOption("lcg.params")[["c"]]
-
-    stopifnot(
-      length(modulus) == 1L, is.numeric(modulus), !is.na(modulus), modulus > 0,
-      length(a) == 1L, is.numeric(a), !is.na(a), a > 0, a < modulus,
-      length(c) == 1L, is.numeric(c), !is.na(c), c >= 0, c < modulus
-    )
-
+  function(seed = NULL) {
     if (!is.null(seed)) {
       if (is.na(seed)) return(.seed)
       stopifnot(length(seed) == 1L, is.numeric(seed), !is.na(seed), seed >= 0)
@@ -66,17 +61,6 @@ lcg <- local({
   }
 })
 
-#' The default LCG parameters are $modulus = 2^16+1$, $a = 75$, and $c = 74$,
-#' which are small enough to to be handled by Bash.  They also happens to be
-#' the one used by the fabolous Sinclair ZX81.
-lcg_set_params <- function(modulus = 2^16+1, a = 75, c = 74) {
-  stopifnot(
-    length(modulus) == 1L, is.numeric(modulus), !is.na(modulus), modulus > 0,
-    length(a) == 1L, is.numeric(a), !is.na(a), a > 0,
-    length(c) == 1L, is.numeric(c), !is.na(c), c > 0
-  )
-  options(lcg.params = c(modulus = modulus, a = a, c = c))
-}
 
 lcg_set_seed <- function(seed) {
   lcg(seed = seed)